Fixed MTL_no_gender_detection.py
