+++ face_attributes_detection/networks.py
@@ -1,0 +1,47 @@
+
+from tensorflow.keras.models import Model
+from tensorflow.keras.layers import Input, Conv2D, BatchNormalization, Activation, MaxPooling2D, Dropout, Flatten, Dense
+
+class FaceNet_NoGender:
+    def __init__(self, shape, channel, unit, first_conv, second_conv):
+        self.shape = shape
+        self.channel = channel
+        self.unit = unit
+        self.first = first_conv
+        self.second = second_conv
+        self.categs = ['mustache', 'eyeglasses', 'beard', 'hat', 'bald']
+
+    def build(self, size, num=2):
+        # initialize the input shape and channel dimension (this code
+        # assumes you are using TensorFlow which utilizes channels
+        # last ordering)
+        inputShape = (self.shape, self.shape, self.channel)
+
+        # construct gender, mustache, bald, eyeglasses, beard & hat sub-networks
+        inputs = Input(shape=inputShape)
+        x = Conv2D(self.first, size, padding="same")(inputs)
+        x = BatchNormalization()(x)
+        x = Activation("relu")(x)
+        x = MaxPooling2D(pool_size=(2, 2))(x)
+        x = Conv2D(self.second, size, padding="same")(x)
+        x = BatchNormalization()(x)
+        x = Activation("relu")(x)
+        x = MaxPooling2D(pool_size=(2, 2))(x)
+
+        flat = Flatten()(x)
+        net = Dense(self.unit)(flat)
+        net = BatchNormalization(axis=-1)(net)
+        net = Activation("relu")(net)
+
+        outputs = []
+        for c in self.categs:
+            tmp_net = Dense(num)(net)
+            tmp_net = Activation('softmax', name=c)(tmp_net)
+            outputs.append(tmp_net)
+
+        # create the model using our input (the batch of images) and
+        # six separate outputs
+        model = Model(inputs=inputs, outputs=outputs, name="facenet")
+
+        # return the constructed network architecture
+        return model
